pyproject.toml
[tool.poetry]
name = "aurora-rsc"
version = "0.1.0"
description = ""
authors = ["twoics <second_vozhzhov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
pydantic-settings = "^2.1.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pytz = "^2024.1"
websockets = "^12.0"
prometheus-fastapi-instrumentator = "^6.1.0"
fastapi-limiter = "^0.1.6"
asgi-correlation-id = "^4.3.1"
pyyaml = "^6.0.1"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
pytest = "^8.0.1"
pytest-asyncio = "^0.23.5"
fakeredis = "^2.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
exclude = [ ".venv" ]
include = [
  "app/config", "app/deps", "app/dto", "app/middleware", "app/models", "app/repo", "app/routes", "app/services", "app/utils", "tests"
]
venvPath = "."
venv = ".venv"
executionEnvironments = [
  { root = "app" },
]
